+++ CHANGELOG.md
@@ -7,7 +7,8 @@ This project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
-- Added `TempDir` for automatically deleted temporary directories.
+- [#7](https://github.com/sunsided/async-tempfile-rs/pull/7):
+  Added `TempDir` for automatically deleted temporary directories.
 
 ## [0.5.0] - 2023-12-06